+++ BUILD
@@ -3015,6 +3015,7 @@ grpc_cc_library(
         "grpc_base",
         "grpc_public_hdrs",
         "grpc_resolver",
+        "grpc_security_base",
         "grpc_service_config_impl",
         "grpc_trace",
         "http_connect_handshaker",
@@ -3038,6 +3039,7 @@ grpc_cc_library(
         "//src/core:channel_stack_type",
         "//src/core:closure",
         "//src/core:construct_destruct",
+        "//src/core:delegating_helper",
         "//src/core:dual_ref_counted",
         "//src/core:env",
         "//src/core:error",
