+++ build_autogenerated.yaml
@@ -833,6 +833,7 @@ libs:
   - src/core/lib/json/json_reader.h
   - src/core/lib/json/json_util.h
   - src/core/lib/json/json_writer.h
+  - src/core/lib/load_balancing/delegating_helper.h
   - src/core/lib/load_balancing/lb_policy.h
   - src/core/lib/load_balancing/lb_policy_factory.h
   - src/core/lib/load_balancing/lb_policy_registry.h
@@ -2214,6 +2215,7 @@ libs:
   - src/core/lib/json/json_object_loader.h
   - src/core/lib/json/json_reader.h
   - src/core/lib/json/json_writer.h
+  - src/core/lib/load_balancing/delegating_helper.h
   - src/core/lib/load_balancing/lb_policy.h
   - src/core/lib/load_balancing/lb_policy_factory.h
   - src/core/lib/load_balancing/lb_policy_registry.h
