+++ tools/doxygen/Doxyfile.c++.internal
@@ -2414,6 +2414,7 @@ src/core/lib/json/json_util.cc \
 src/core/lib/json/json_util.h \
 src/core/lib/json/json_writer.cc \
 src/core/lib/json/json_writer.h \
+src/core/lib/load_balancing/delegating_helper.h \
 src/core/lib/load_balancing/lb_policy.cc \
 src/core/lib/load_balancing/lb_policy.h \
 src/core/lib/load_balancing/lb_policy_factory.h \
